Improve thread safety
Move shared data into a single block, to help reason about what needs to
be protected by the mutex, as both the task queue and the active boolean
need to be shared.

Also clean up the worker loop a little to help readability. The lock can
exist throughout the worker loop, as it is released only when running
the task or waiting. This ensures that the behaviour is as expected with
as little code as needed.