Add basic Tuple implementations (#2)
Provide fairly minimal implementations of Tuple, using three different
techniques. The inheritance based method allows easier access to
elements and potentially packs elements better, while the member based
implementation will be standard layout and so could be used where that
is required. The flat inheritance tuple is similar to that used in libc++ where
all elements are inherited at the base level which ensure the compiler can
organise the layout as it sees fit, but requires more complex indexing.